% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_mem_usage.R
\name{check_mem_usage}
\alias{check_mem_usage}
\title{Check memory usage. Stop programm if memory usage is to hight.}
\usage{
check_mem_usage(limit = 2000)
}
\arguments{
\item{limit}{numeric. Define minimum free memory in MB}
}
\description{
Check memory usage. Stop programm if memory usage is to hight.
}
\examples{
check_mem_usage()
}
