% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_pts.R
\name{create_pts}
\alias{create_pts}
\title{Create points along SpatialLinesDataFrame}
\usage{
create_pts(
  l,
  id_l = "ID_LINES",
  id_p = "ID_PTS",
  step = 5,
  ncores = NULL,
  p = T
)
}
\arguments{
\item{l}{a SpatialLinesDataFrame}

\item{id_l}{default NULL. A character string. If NULL, row.names is use. If set, it define column name used.}

\item{id_p}{default "ID_PTS". A character string. If set, it define column name of unique points IDs.}

\item{step}{defautl 5. An integer. Define the distance between two points along the line in meters.}

\item{ncores}{default NULL. An integer. If NULL all CPU cores - 1 are used. If set, define number of CPU cores used.}

\item{p}{default TRUE. Boolean. Define if function use multithreading.}
}
\value{
a SpatialPointsDataFrame
}
\description{
Create points along SpatialLinesDataFrame
}
\examples{

# return lol
}
