% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cut_sldf.R
\name{cut_sldf}
\alias{cut_sldf}
\title{Cut a SpatialLinesDateFrame into a smaller segments}
\usage{
cut_sldf(l, id_l)
}
\arguments{
\item{l}{a SpatialLinesDataFrame.}

\item{id_l}{a character string. IDs Column name of SpatialLinesDataFrame.}
}
\value{
a SpatialLinesDataFrame
}
\description{
Cut a SpatialLinesDateFrame into a smaller segments
}
