% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_sldf.R
\name{filter_sldf}
\alias{filter_sldf}
\title{Title}
\usage{
filter_sldf(l_A, l_B, id_l_B, bfr_width)
}
\arguments{
\item{l_A}{a SpatialLinesDataFrame. Reference lines.}

\item{l_B}{a SpatialLinesDataFrame. Searched lines.}

\item{id_l_B}{a character string. IDs Column name of searched SpatialLinesDataFrame (Lines B)}

\item{bfr_width}{an integer. Buffer distance.}
}
\value{
a SpatialLinesDataFrame.
}
\description{
Title
}
