% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_np.R
\name{find_np}
\alias{find_np}
\title{Find nearest point between two SpatialPointsDataFrame.}
\usage{
find_np(pts_A, pts_B, id_p_A, id_p_B = "ID_PTS_B", ncores = NULL, p = T)
}
\arguments{
\item{pts_A}{a SpatialPointsDataFrame. Reference points.}

\item{pts_B}{a SpatialPointsDataFrame. Searched points.}

\item{id_p_A}{a character string. Indicates the IDs column name of pts_A.}

\item{id_p_B}{a character string. Default "ID_PTS_B". Indicates the IDs column name of pts_B.}

\item{ncores}{an integer. Number of CPU cores used. If NULL, all cores - 1 are used.}

\item{p}{logical. Define if function use multithreading. Default = TRUE.}
}
\value{
a data.table
}
\description{
Find nearest point between two SpatialPointsDataFrame.
}
\examples{
#test
}
