% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nnl.R
\name{nnl}
\alias{nnl}
\title{Complete and unique function to extrapolate a SpatialLinesDataFrame A to a SpatialLinesDataFrame B}
\usage{
nnl(
  l_A,
  l_B,
  id_l_A,
  id_l_B,
  id_p_A = "ID_PTS_A",
  id_p_B = "ID_PTS_B",
  step = 5,
  rate = 45,
  p = T,
  ncores = NULL
)
}
\arguments{
\item{l_A}{a SpatialLinesDataFrame. Reference lines.}

\item{l_B}{a SpatialLinesDataFrame. Searched lines.}

\item{id_l_A}{a character string. IDs Column name of reference SpatialLinesDataFrame (Lines A)}

\item{id_l_B}{a character string. IDs Column name of reference SpatialLinesDataFrame (Lines B)}

\item{id_p_A}{a character string. Indicates the IDs column name of points created to lines A. Default : "ID_PTS_A"}

\item{id_p_B}{a character string. Indicates the IDs column name of points created to lines B. Default : "ID_PTS_B"}

\item{step}{an integer. Define the distance between two points along the line in meters. Defautl = 5.}

\item{rate}{an integer. Between 0 and 100. Define lower limit to select line B. Rate between number of points B on Lines A and totalpoints on Line B.}

\item{p}{a boolean. Define if function use multithreading. Default = TRUE.}

\item{ncores}{an integer. Number of CPU cores used. If NULL, all cores - 1 are used.}
}
\value{
a SpatialLinesDataFrame.
}
\description{
Complete and unique function to extrapolate a SpatialLinesDataFrame A to a SpatialLinesDataFrame B
}
