% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nnl_step1.R
\name{nnl_step1}
\alias{nnl_step1}
\title{Find near neighbors lines. First step which can contains discontinuities.}
\usage{
nnl_step1(pts_A, pts_B, fnp_A, id_p_A, id_p_B, id_l_A, id_l_B, rate = 45)
}
\arguments{
\item{pts_A}{a SpatialPointsDataFrame. Reference points.}

\item{pts_B}{a SpatialPointsDataFrame. Searched points.}

\item{fnp_A}{a data.table. Result of find_np().}

\item{id_p_A}{a character string. Indicates the IDs column name of pts_A.}

\item{id_p_B}{a character string. Indicates the IDs column name of pts_B.}

\item{id_l_A}{a character string. IDs Column name of reference SpatialLinesDataFrame (Lines A)}

\item{id_l_B}{a character string. IDs Column name of searched SpatialLinesDataFrame (Lines A)}

\item{rate}{an integer. Between 0 and 100. Define lower limit to select line B. Rate between number of points B on Lines A and totalpoints on Line B.}
}
\value{
a data.frame. Contains near neighbors lines
}
\description{
Find near neighbors lines. First step which can contains discontinuities.
}
