% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nnl_step2.R
\name{nnl_step2}
\alias{nnl_step2}
\title{Find discontinuities in nnl_step1() result}
\usage{
nnl_step2(l_B, r_s1_A, id_l_A, id_l_B)
}
\arguments{
\item{l_B}{a SpatialLinesDataFrame. Searched lines.}

\item{r_s1_A}{a data.frame. Result of nnl_step1().}

\item{id_l_A}{a character string. IDs Column name of reference SpatialLinesDataFrame (Lines A)}

\item{id_l_B}{a character string. IDs Column name of searched SpatialLinesDataFrame (Lines B)}
}
\value{
a data.frame. Contains discontinuities.
}
\description{
Find discontinuities in nnl_step1() result
}
